Serializer for the GraphUpdateStack.
This will be needed if we want to save and retrieve the state of
feature computation, and important for features that take long to
compute.

For instance when we compute the mean spot intensity (which takes very
long), as long as we do not save the feature update mechanism works.
We only need to recompute the mean for spots that changes since the last
feature computation.

So if we want to keep this nicety after deserialization of the feature
we must also deserialize the state of the feature update.